--- src/pages/Dash.tsx
@@ -1,194 +1,0 @@
-import {
-  Avatar,
-  Box,
-  Drawer,
-  DrawerContent,
-  DrawerOverlay,
-  Flex,
-  Icon,
-  IconButton,
-  Image,
-  Input,
-  InputGroup,
-  InputLeftElement,
-  Menu,
-  MenuButton,
-  MenuDivider,
-  MenuGroup,
-  MenuItem,
-  MenuList,
-  useColorModeValue,
-  useDisclosure,
-} from "@chakra-ui/react";
-import React from "react";
-import { FaBell } from "react-icons/fa";
-import { FiMenu, FiSearch } from "react-icons/fi";
-import ThemeSwitch from "../components/ThemeSwitch";
-import {
-  LOGO_DARK,
-  LOGO_WHITE,
-  SIDEBAR_THEME_DARK,
-  SIDEBAR_THEME_LIGHT,
-} from "../constants/appConstants";
-import { useAuth } from "../contexts/AuthProvider";
-import NavbarMenu from "../utils/navbarMenu";
-
-export default function Dash() {
-  const sidebar = useDisclosure();
-
-  const NavItem = (props) => {
-    const { icon, children, ...rest } = props;
-    return (
-      <Flex
-        align="center"
-        px="4"
-        mx="2"
-        rounded="md"
-        py="3"
-        cursor="pointer"
-        color={useColorModeValue("blackAlpha.800", "whiteAlpha.700")}
-        _hover={{
-          bg: "blackAlpha.300",
-          color: "whiteAlpha.900",
-        }}
-        role="group"
-        fontWeight="semibold"
-        transition=".15s ease"
-        {...rest}
-      >
-        {icon && (
-          <Icon
-            mr="5"
-            boxSize="4"
-            _groupHover={{
-              color: "gray.300",
-            }}
-            as={icon}
-          />
-        )}
-        {children}
-      </Flex>
-    );
-  };
-
-  const SidebarContent = (props) => (
-    <Box
-      as="nav"
-      pos="fixed"
-      top="0"
-      left="0"
-      zIndex="sticky"
-      h="full"
-      pb="10"
-      overflowX="hidden"
-      overflowY="auto"
-      bg={useColorModeValue(SIDEBAR_THEME_LIGHT, SIDEBAR_THEME_DARK)}
-      borderColor="blackAlpha.300"
-      borderRightWidth="1px"
-      w="60"
-      {...props}
-    >
-      <Flex px="4" py="5" align="center">
-        <Image src={useColorModeValue(LOGO_WHITE, LOGO_DARK)} />
-      </Flex>{" "}
-      <Flex
-        direction="column"
-        as="nav"
-        fontSize="sm"
-        color="gray.600"
-        aria-label="Main Navigation"
-      >
-        {NavbarMenu.map(({ name, icon, url, key, ...rest }) => (
-          <NavItem icon={icon} key={key} {...rest}>
-            {name}
-          </NavItem>
-        ))}
-      </Flex>
-    </Box>
-  );
-  const { signOut } = useAuth();
-  return (
-    <Box
-      as="section"
-      bg={useColorModeValue("gray.50", "gray.700")}
-      minH="100vh"
-    >
-      <SidebarContent display={{ base: "none", md: "unset" }} />
-      <Drawer
-        isOpen={sidebar.isOpen}
-        onClose={sidebar.onClose}
-        placement="left"
-      >
-        <DrawerOverlay />
-        <DrawerContent>
-          <SidebarContent w="full" borderRight="none" />
-        </DrawerContent>
-      </Drawer>
-      <Box ml={{ base: 0, md: 60 }} transition=".3s ease">
-        <Flex
-          as="header"
-          align="center"
-          justify="space-between"
-          w="full"
-          px="4"
-          py="5"
-          bg={useColorModeValue("white", "gray.800")}
-          borderBottomWidth="1px"
-          borderColor="blackAlpha.300"
-        >
-          <IconButton
-            aria-label="Menu"
-            display={{ base: "inline-flex", md: "none" }}
-            onClick={sidebar.onOpen}
-            icon={<FiMenu />}
-            size="sm"
-          />
-          <InputGroup w="96" display={{ base: "none", md: "flex" }}>
-            <InputLeftElement color="gray.500" children={<FiSearch />} />
-            <Input placeholder="Search for articles..." />
-          </InputGroup>
-
-          <Flex align="center">
-            <ThemeSwitch />
-            <IconButton
-              bg="transparent"
-              aria-label="notifications"
-              icon={<FaBell />}
-              onClick={() => {}}
-            />
-            <Menu>
-              <MenuButton
-                as={Avatar}
-                ml="4"
-                size="sm"
-                name="piyush97"
-                src="https://avatars.githubusercontent.com/piyush97"
-                cursor="pointer"
-              ></MenuButton>
-              <MenuList>
-                <MenuGroup title="Profile">
-                  <MenuItem>My Account</MenuItem>
-                  <MenuItem>Payments </MenuItem>
-                </MenuGroup>
-                <MenuDivider />
-                <MenuGroup title="Help">
-                  <MenuItem>Docs</MenuItem>
-                  <MenuItem>FAQ</MenuItem>
-                </MenuGroup>
-                <MenuDivider />
-                <MenuItem cursor="pointer" onClick={() => signOut()}>
-                  Log Out
-                </MenuItem>
-              </MenuList>
-            </Menu>
-          </Flex>
-        </Flex>
-
-        <Box as="main" p="4">
-          {/* Add content here, remove div below  */}
-          <Box borderWidth="4px" borderStyle="dashed" rounded="md" h="96" />
-        </Box>
-      </Box>
-    </Box>
-  );
-}
