+++ src/pages/dashboard.tsx
@@ -1,12 +1,12 @@
 import { RouteComponentProps } from "@reach/router";
 import React from "react";
+import DashboardContainer from "../containers/DashboardContainer";
 import { useAuth } from "../contexts/AuthProvider";
-import Dash from "./Dash";
 import Login from "./Login";
 
 const Dashboard = (props: RouteComponentProps) => {
   const { isAuth } = useAuth();
-  return isAuth ? <Dash /> : <Login />;
+  return isAuth ? <DashboardContainer /> : <Login />;
 };
 
 export default Dashboard;
